add show all details

+++ AUOpenHouseStudentIOS/Models/Faculty+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Faculty+CoreDataProperties.swift
 //  AUOpenHouseStudentIOS
 //
-//  Created by ios-project on 9/6/2561 BE.
+//  Created by ios-project on 10/6/2561 BE.
 //
 //
 
@@ -19,8 +19,8 @@ extension Faculty {
     @NSManaged public var fid: Int32
     @NSManaged public var icon: URL?
     @NSManaged public var info: String?
-    @NSManaged public var location_latitude: Float
-    @NSManaged public var location_longitude: Float
+    @NSManaged public var location_latitude: Double
+    @NSManaged public var location_longitude: Double
     @NSManaged public var name: String?
     @NSManaged public var website: URL?
     @NSManaged public var majors: NSSet?
